Automatically compile shaders to temporary directory.
This commit adds initial built-in shader support to the meson build system.
Since all meson generated files are located within the temporary build
directory (as object files are) and this behaviour cannot be changed,
it is required to remove the "Shaders/" prefix from the #includes in
RealDevice.cpp.

In order to use the same source files without modification for both
build systems, the Visual Studio build system now writes the compiled
shaders to the temporary build directory as well.