Implement LOOP, ENDLOOP and BREAK DXBC ops
We also refactor the REP and ENDREP ops to share code with the LOOP AND
ENDLOOP ops. LOOP and ENDLOOP are only partially implemented because
loads and (unlikely) stores to the counter variable are not yet
supported.

A future refactoring is expected to support loads from the counter. The
plan is to add a stack that would be referenced whenever an attempt to
access the aL variable is made. I had wanted to do testing before
implementing that, but VK9's dependence on boost has kept me from cross
compiling it in the time that I could spare, delaying testing.

Chris asked me to put what I had ready in a pull request anyway, so I am
doing that.

Signed-off-by: Richard Yao <ryao@gentoo.org>